requirements.txt
python-telegram-bot==21.6
google-genai>=0.5.0
python-dotenv>=1.0.1